requirements.txt
flask
python-dotenv
paho-mqtt>=1.6.1